Tile Based Learning
Changed the background algorithm to take in the 4 tiles in which it can move, aswell as the full map, so that it has better context of its decision.

Added reward for hurting enemy
Added method to visualize Q values for the last state at the end of a run

Adjusted penalty for revisiting tiles.
Attempt to fix Mass Equip bug